+++ CHANGELOG.md
@@ -6,6 +6,10 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 ## [Unreleased]
 
+### Fixed
+
+- Dash app button visibility height.
+
 ## [21.0.0 Reynolds] - 2022-08-06
 
 ### Changed