+++ CHANGELOG.md
@@ -6,14 +6,17 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 ## [Unreleased]
 
+## [22.0.0 Millet] - 2022-09-10
+
 ### Fixed
 
 - Dash app button visibility height.
-- Looking glass recursion.
+- Looking glass error after unlock.
 
 ### Added
 
 - App menu label visibility.
+- GNOME Shell 43 support.
 - Quick settings menu visibility.
 
 ### Removed